Import puller.sh from mobileapps
Imports puller.sh, a script used to automatically update and restart the
service when set up to run in a standalone labs instance. This is needed
for setting up a standalone instance for testing.

N.B. This is the original version of the puller script, without the recent
hack to use a more up-to-date npm version via nvm. That change will be
reverted when mobileapps is running in an instance with a more up-to-date
packaged npm version.

Change-Id: I3657d2f68bacb7f76d045704d632271704968b19