Fix: Use templated request for MediaWiki page requests
Eliminates a raw preq request to a public Wikipedia URL in the course of
generating the /page/featured response. It's replaced with a templated
request following the pattern currently used for MW API and RESTBase
requests. This fixes the issue that necessitated rollback of the service
deployment and allows the service to be deployed in Kubernetes.

Bug: T170455
Change-Id: Idda12664619f70200eb3961fd61a9e2e01c3b450