Fix: Use localized title when requesting Special:FeedItem
Many non-English wikis have localized aliases for the Special namespace
and the FeedItem special page title. Typically, a request for the
default 'Special:FeedItem' page redirects to the localized version.
However, since the redirect locations are provided as external
URLs, they cannot be followed when making requests strictly internal to
the production cluster.

To work around this, we should localize the title used in the special
page request. This can be done with info already being requested from
siteinfo.

Bug: T235102
Change-Id: Ideaf5b17dc5099306c0970d9bebc2a363141d0cb