Update to node 10 and bump dependency versions
Use the node 10 base images in the pipeline. This better conforms to our
current development environment and allows us to update dependencies to
latest versions.

With this change, all potential vulnerabilities currently identified by npm
audit are resolved.

Some eslint-ignore lines are removed as a result of the updates.

Hygiene: The 'deploy' key is removed from package.json. This was only used
to define parameters for service-runner-based Docker builds, and is no
longer needed.

Change-Id: Ic0d952b25180d581ae2645f860c38cec939b4594