Throw 404 and not 501 for endpoints unsupported for the request domain
Currently the service throws very frequent 501 unsupported_language
errors. Indeed these appear to make up the bulk of error responses
reported in production. A 5xx should not be thrown in this case because
the server is not actually in an error state. It is expected for the
news, featured article, and on-this-day responses not to be implemented
for most languages. Further, a 501 response is not semantically correct
per RFC 7231; a 501 is appropriate "when the server does not recognize the
request method and is not capable of supporting it for any resource."

https://tools.ietf.org/html/rfc7231#section-6.6.2

A 404 is appropriate in this situation, and should be thrown instead.

Change-Id: Icd0d091baba5461b3eaac2ccb032d7a203d48d6f