Hygiene: Move over random page title endpoint from mobileapps
The random page title endpoint is intended to support the Explore feed
in the apps, and so should live here rather than in mobileapps.

The response is updated from the mobileapps version to stop sending an
ETag. Sending an ETag isn't necessary, as the random page response isn't
intended to be cached.

The endpoint minor version is bumped from 0.6.0 in mobileapps to 0.7.0
here, to reflect dropping the ETag, as well as a recent change to stop
requesting text extracts from the MW API to break ties.

Change-Id: I1979b4312b8ba1f3fef38933991bdae3e778975f