Add survey announcement for iOS
iOS needs to lean on the announcements system to display a new survey prompt when they land on an
article. We need additional info just for this announcement (articleTitles and displayDelay). To
handle this cleanly, I refactored much of the config variables into campaign objects. With this
hopefully it will be easier in the future to use this announcements system with multiple campaigns
running at the same time (for example, a survey and a fundraising campaign on overlapping time
periods).

Bug: T248507
Change-Id: If997d1def0f1f6ff04b813bdba1be5f688205302