Random: Rely on pageprops to determine whether a page image exists
Currently the MW API request for /page/random requests
prop=pageimages to determine whether a page has a page image. We
should use pageprops instead because it involves only a simple and
fast DB query, unlike pageimages, which may involve expensive (and in
this case totally unnecessary) thumbnail processing.

If the title has a pageprops entry for page_image or page_image_free,
it has a page image.

Bug: T260638
Change-Id: Ic87ff26c9432632d896e502b7ee6e734fa6ce5a3