Update iOS announcement survey object for next iteration
We will soon run a new iteration of our survey to collect values against our upcoming
article as a living document experiment. Since the announcement object already served
as a gatekeeper/allowlist for the survey, we are adding a percent_receiving_experiment
field to give us a scalable ratio on which bucket of the experiment the user will fall
into (test or control). This will give us the ability to turn the experiment off server-side,
or back off if our labs instance needs less traffic. We are also expanding the article list
to additional top read articles.

Bug: T266274
Change-Id: I30fc3db61feb424bce0c9692cb46082ff23ec1c7