Announcements: Raise max-age, lower s-maxage
Updates the cache control header for announcments to allow clients to
cache responses for up to 24 hours, and shared caches (e.g., Varnish)
to retain them for up to 15 minutes.

Lowering the s-maxage will make changes to announcements available to
clients sooner without manual intervention going forward. Also, the
announcements response is cheap to generate, so it should not have a
signficant impact on service operation.

Change-Id: If1a44b9ccd8d7e5213ed05687380fb5b6d7a3f5a