Explicitly add country to announcements URL
The apps know what country a device is in, and show the appropriate fundraising message for a given country. If a user clicks on the link, our donate subdomain then shows different options based on a user's country. This patch explicitly tells the website which country to show. Previously, the donate domain was also trying to determine which country a user is in, and if certain privacy settings were on the mobile browser it was unable to do so.

Change-Id: Ibedb27f11dc7c62d1e720413de90b526d8dcaaa9